pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "configdiscovery"
version = "0.1.0"
description = "LLM-driven HPC configuration discovery and capture via Globus Compute"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "anthropic>=0.40.0",
    "globus-compute-sdk>=2.0.0",
    "PyGithub>=2.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
configdiscovery = "configdiscovery.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["configdiscovery"]
